raft: move progress to progress.go
